+++ src/lib/initials.js
@@ -1,0 +1,9 @@
+var hepburn = require('hepburn');
+
+function initials(text) {
+  return hepburn
+  .fromKana(text.toUpperCase())
+  .substring(0, 2);
+}
+
+module.exports = initials;
